requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
Babel==2.12.1
billiard==3.6.4.0
black==23.1.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap5==0.7
cron-descriptor==1.2.35
cryptography==39.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1.7
django-celery-beat==2.5.0
django-crispy-forms==2.0
django-extensions==3.2.1
django-filter==23.1
django-location-field==2.1.0
django-money==3.0.0
django-phonenumber-field==7.0.2
django-templated-mail==1.1.1
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djongo==1.3.6
djoser==2.1.0
dnspython==2.3.0
drf-yasg==1.21.5
Faker==17.3.0
flake8==6.0.0
flower==1.2.0
graphviz==0.20.1
gunicorn==20.1.0
humanize==4.6.0
idna==3.4
inflection==0.5.1
install==1.3.5
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.0
pathspec==0.11.0
phonenumbers==8.13.6
Pillow==9.4.0
platformdirs==3.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
py-moneyed==2.0
pycodestyle==2.10.0
pycparser==2.21
pydot==1.4.2
pydot-ng==2.0.0
pyflakes==3.0.1
PyJWT==2.6.0
pymongo==4.3.3
pyparsing==3.0.9
python-crontab==2.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7.1
redis==4.3.4
requests==2.28.2
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.2.4
tomli==2.0.1
tornado==6.2
typing_extensions==4.5.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
webargs==8.2.0
wrapt==1.15.0
